fix tests to isolate the ExifTool and ExifToolHelper tests changes the way tests are run against both the base class and the helper class
